rename extensions -> std_extensions
